+++ src/App.tsx
@@ -5,7 +5,7 @@ import { EmailConfirmed } from './pages/email-confirmed';
 import { ForgotPassword } from './pages/forgot-password';
 import { Profile } from './pages/profile';
 import { Room } from './pages/room';
-import { UserSearch } from './pages/user-search';
+import { Explore } from './pages/explore';
 import { ThemeProvider } from './components/theme-provider';
 import { Toaster } from 'sonner';
 
@@ -35,7 +35,7 @@ export default function App() {
                 <Route path="/forgot-password" element={<ForgotPassword />} />
                 <Route path="/profile/:displayedUserId" element={<Profile />} />
                 <Route path="/room" element={<Room />} />
-                <Route path="/user-search" element={<UserSearch />} />
+                <Route path="/explore" element={<Explore />} />
             </Routes>
         </ThemeProvider>
     );
