+++ src/components/nav-bar-footer.tsx
@@ -31,7 +31,7 @@ export const NavBarFooter = () => {
                             {
                                 icon: SearchIcon,
                                 label: "Explore",
-                                link: "/user-search"
+                                link: "/explore"
                             }
                         ].map((option) => (
                             <Link to={option.link} key={option.link}>
